Refactor the common parts of the implementations
